refactor(consumer): move to consumer package
