+++ consumer/consumer_integration_test.go
@@ -1,4 +1,4 @@
-package gorabbit_test
+package consumer_test
 
 import (
 	"context"
@@ -6,12 +6,12 @@ import (
 	"time"
 
 	"github.com/heureka/gorabbit/channel"
+	"github.com/heureka/gorabbit/consumer"
 	amqp "github.com/rabbitmq/amqp091-go"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/heureka/gorabbit"
 	"github.com/heureka/gorabbit/rabbittest"
 )
 
@@ -22,7 +22,7 @@ type ConsumerTestSuite struct {
 //nolint:gocognit // complexity is fine for a test
 func (s *ConsumerTestSuite) TestConsume() {
 	tests := map[string]struct {
-		options      []gorabbit.Option
+		options      []consumer.Option
 		messages     [][]byte
 		consumeErr   error
 		wantReceived [][]byte
@@ -47,10 +47,10 @@ func (s *ConsumerTestSuite) TestConsume() {
 		s.Run(name, func() {
 			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 			defer cancel()
-			ch, err := channel.New(s.Connection)
+			ch, err := channel.New(s.Connection.Connection)
 			s.Require().NoError(err, "create channel")
 
-			txreader := gorabbit.NewConsumer(
+			cons := consumer.New(
 				ch,
 				s.Queue,
 				tt.options...,
@@ -59,15 +59,15 @@ func (s *ConsumerTestSuite) TestConsume() {
 			receiveCh := make(chan []byte)
 			defer close(receiveCh)
 
-			consumer := newMockEchoConsumer(receiveCh)
-			consumer.On("Process", mock.Anything, mock.Anything).Return(tt.consumeErr)
+			processor := newMockEchoProcessor(receiveCh)
+			processor.On("Process", mock.Anything, mock.Anything).Return(tt.consumeErr)
 
 			var eg errgroup.Group
 			eg.Go(func() error {
-				return txreader.Start(ctx, consumer)
+				return cons.Start(ctx, processor)
 			})
 
-			// publish all messages
+			// publisher all messages
 			for _, m := range tt.messages {
 				s.publish(ctx, m)
 			}
@@ -78,7 +78,7 @@ func (s *ConsumerTestSuite) TestConsume() {
 				received = append(received, <-receiveCh)
 			}
 
-			s.Require().NoError(txreader.Stop(), "must not return error on stop")
+			s.Require().NoError(cons.Stop(), "must not return error on stop")
 			s.Require().NoError(eg.Wait(), "must not return error on starting")
 
 			s.Assert().ElementsMatch(tt.wantReceived, received, "should consume expected messages")
@@ -87,15 +87,12 @@ func (s *ConsumerTestSuite) TestConsume() {
 }
 
 func (s *ConsumerTestSuite) TestImmediatelyStop() {
-	ch, err := channel.New(s.Connection)
+	ch, err := channel.New(s.Connection.Connection)
 	s.Require().NoError(err)
 
-	rmq := gorabbit.NewConsumer(ch, s.Queue)
-	if err != nil {
-		s.FailNow("create rmq", err)
-	}
+	cons := consumer.New(ch, s.Queue)
 
-	err = rmq.Stop()
+	err = cons.Stop()
 	s.Assert().NoError(err, "should not return error")
 }
 
@@ -122,7 +119,7 @@ type mockEchoConsumer struct {
 	received chan<- []byte
 }
 
-func newMockEchoConsumer(received chan<- []byte) *mockEchoConsumer {
+func newMockEchoProcessor(received chan<- []byte) *mockEchoConsumer {
 	return &mockEchoConsumer{
 		received: received,
 	}
